requirement.txt
streamlit 
opencv-python-headless 
numpy 
pillow 
exifread 
pandas 
matplotlib